Firgured it out
Rather then tracking the index in one function and removing them in an other I decided to do it all in one function.

I also found the problem with the repeated '%'. Simply put I was not changing the counter correctly. So if there was a repeated '%' I would pass it over because it was in a lower index then my counter.